requirements.txt.txt
fastapi
uvicorn
python-dotenv
apscheduler
httpx